image_classifier/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
dramatiq==1.17.1
filelock==3.16.1
fsspec==2024.10.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
pillow==11.0.0
prometheus_client==0.21.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
